Move function to shared place
